requirements.txt
pytest>=7.2.1
selene>=2.0.0b16
allure-pytest>=2.12.0
requests==2.28.2
voluptuous==0.13.1
pytest-voluptuous==1.2.0
python-dotenv==0.21.0
curlify